% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendationengine.R
\name{recommendation_engine}
\alias{recommendation_engine}
\title{Recommendation Engine Function}
\usage{
recommendation_engine(data, top)
}
\arguments{
\item{data}{provide the data frame with customer id and historical product purchase}
}
\description{
This function allows you to build item based recommendation engine for customers
}
\examples{
recommendation_engine()
}
\keyword{recommendation}

